requirements.txt
easyocr==1.7.0
ultralytics==8.0.200
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless==4.8.1.78
Pillow==10.0.0
pdf2image==1.16.3
rapidfuzz==3.5.2
numpy==1.24.3
pandas==2.0.3
